feat: huge refactoring
