+++ .travis.yml
@@ -7,15 +7,13 @@ branches:
     - master
 install:
   - npm install
-  - npm install -g snyk
-  - snyk test
 cache:
   directories:
     - node_modules
 script:
-  - NODE_ENV=test npm run-script test-travis
+  - npm run lint && npm run test:cov
 # Send coverage data to Coveralls
-after_script: "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
+after_script: "npm run coverage"
 after_success:
   - |
     if ([ "$TRAVIS_BRANCH" == "master" ] || [ ! -z "$TRAVIS_TAG" ]) && [ "$TRAVIS_PULL_REQUEST" == "false" ]; then
@@ -27,9 +25,8 @@ after_success:
       npm version patch -m "$(git log -1 --pretty=%B) .... bump version [skip ci]"
       git push origin version-branch:master --follow-tags
     else
-      echo "version skiped!"
+      echo "version skipped!"
     fi
-  - snyk monitor
 deploy:
   skip_cleanup: true
   provider: npm
