+++ lib/const/inputs-to-relations.js
@@ -1,8 +1,8 @@
 const RELATIONS = {
-    "@": 'waitNode',
-    "*@": 'waitAny',
-    "#@": 'waitBatch',
-    "*#@": 'waitAnyBatch'
+    '@': 'waitNode',
+    '*@': 'waitAny',
+    '#@': 'waitBatch',
+    '*#@': 'waitAnyBatch'
 };
 
 module.exports = RELATIONS;
